+++ src/jsMain/kotlin/at/posselt/pfrpg/actor/PF2EParty.kt
@@ -1,0 +1,5 @@
+package at.posselt.pfrpg.actor
+
+import com.foundryvtt.core.documents.Scene
+import com.foundryvtt.pf2e.actor.PF2EParty
+