+++ src/jsMain/kotlin/at/posselt/pfrpg/migrations/migrations/Migration11.kt
@@ -7,7 +7,6 @@ import com.foundryvtt.core.Game
 
 class Migration11 : Migration(11) {
     override suspend fun migrateCamping(game: Game, camping: CampingData) {
-
     }
 
 }